MXS-2146: Add csmon test case
Added a test case that does a set of sanity checks on the monitor. As the
monitor is very simple, there are not a lot of things to test without
access to the actual instances (e.g. ExeMgr failures need to be tested).

Currently the test always passes as ColumnStore clusters aren't
implemented for the test framework.